refactor(TabletsStatistic): migrate to ts
